### 2023-10-01 #5913 @editorialbot
[PRE REVIEW]: cppdlr: Imaginary time calculations using the discrete Lehmann representation
**Submitting author:** @jasonkaye (<a href="http://orcid.org/0000-0001-8045-6179">Jason Kaye</a>)
**Repository:** https://github.com/flatironinstitute/cppdlr/
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
