### 2023-10-03 #5913 @HugoStrand
Dear Editor,

In your list of reviewers `Evan1415` https://reviewers.joss.theoj.org/reviewers/815 seems to have background in the field.

Best regards, Hugo U. R. Strand (co-author)
