### 2023-10-04 94e74d1
Fix issue with non-contiguous views in vals2coefs implementations
