### 2023-10-04 3f53659
Fix issue with non-contiguous views in vals2coefs implementations
