### 2023-10-06 a8693fb
added program to print ranks
