### 2023-10-06 b48045d
added l2 err, im freq errs to im time test
