### 2023-10-06 9eb498a
update tests to measure l2 error
