### 2023-10-06 be35cda
bug fix geterr_k_it
