### 2023-10-06 4ce9da9
implemented barycentric legendre interpolation
