### 2023-10-06 6509021
updated k matrix accuracy test for legendre discretization
