### 2023-10-06 6df4d6d
clean up of l2 weighting modifications
