+++ docs/main/doxygen/latex/refman.tex
@@ -181,6 +181,7 @@
 \input{namespacecppdlr}
 \chapter{Class Documentation}
 \input{classcppdlr_1_1barycheb}
+\input{classcppdlr_1_1baryleg}
 \input{classdyson__it}
 \input{classcppdlr_1_1fineparams}
 \input{classcppdlr_1_1imfreq__ops}
