+++ docs/main/doxygen/latex/classcppdlr_1_1barycheb.tex
@@ -11,20 +11,20 @@
 \item 
 nda\+::vector$<$ double $>$ const  \& \mbox{\hyperlink{classcppdlr_1_1barycheb_afac15e0f486dc87d3f550a5b385117de}{getnodes}} ()
 \item 
-double \mbox{\hyperlink{classcppdlr_1_1barycheb_a3be85efddb7f3d2a0432f04a4529eda2}{interp}} (double x, nda\+::vector\+\_\+const\+\_\+view$<$ double $>$ f)
+double \mbox{\hyperlink{classcppdlr_1_1barycheb_ab2f5044a724a2450499000f078b495c3}{interp}} (double xeval, nda\+::vector\+\_\+const\+\_\+view$<$ double $>$ f)
 \end{DoxyCompactItemize}
 
 
 \doxysubsection{Detailed Description}
-Initialize subroutine barycheb for barycentric Lagrange interpolation at Chebyshev nodes.
+Class constructor for barycheb\+: barycentric Lagrange interpolation at Chebyshev nodes.
 
 
 \begin{DoxyParams}{Parameters}
 {\em n} & number of Chebyshev nodes \\
 \hline
-{\em xc} & n Chebyshev nodes of the first kind \\
+{\em x} & n Chebyshev nodes of the first kind \\
 \hline
-{\em wc} & barycentric interpolation weights at Chebyshev nodes of the first kind \\
+{\em w} & barycentric interpolation weights at Chebyshev nodes of the first kind \\
 \hline
 \end{DoxyParams}
 
@@ -45,11 +45,11 @@ Initialize subroutine barycheb for barycentric Lagrange interpolation at Chebysh
 \doxysubsubsection{\texorpdfstring{getnodes()}{getnodes()}}
 {\footnotesize\ttfamily nda\+::vector$<$ double $>$ const  \& cppdlr\+::barycheb\+::getnodes (\begin{DoxyParamCaption}{ }\end{DoxyParamCaption})}
 
-\mbox{\Hypertarget{classcppdlr_1_1barycheb_a3be85efddb7f3d2a0432f04a4529eda2}\label{classcppdlr_1_1barycheb_a3be85efddb7f3d2a0432f04a4529eda2}} 
+\mbox{\Hypertarget{classcppdlr_1_1barycheb_ab2f5044a724a2450499000f078b495c3}\label{classcppdlr_1_1barycheb_ab2f5044a724a2450499000f078b495c3}} 
 \index{cppdlr::barycheb@{cppdlr::barycheb}!interp@{interp}}
 \index{interp@{interp}!cppdlr::barycheb@{cppdlr::barycheb}}
 \doxysubsubsection{\texorpdfstring{interp()}{interp()}}
-{\footnotesize\ttfamily double cppdlr\+::barycheb\+::interp (\begin{DoxyParamCaption}\item[{double}]{x,  }\item[{nda\+::vector\+\_\+const\+\_\+view$<$ double $>$}]{f }\end{DoxyParamCaption})}
+{\footnotesize\ttfamily double cppdlr\+::barycheb\+::interp (\begin{DoxyParamCaption}\item[{double}]{xeval,  }\item[{nda\+::vector\+\_\+const\+\_\+view$<$ double $>$}]{f }\end{DoxyParamCaption})}
 
 
 
