### 2023-10-06 1f632c7
tightened down some test tolerances
