### 2023-10-27 a7162f5
increased test tolerance due to slight failure for gcc 13.2
