+++ test/c++/imfreq_ops.cpp
@@ -271,7 +271,7 @@ TEST(imfreq_ops, interp_matrix_sym_fer) {
   auto gc = ifops.vals2coefs(beta, g);
 
   // Check that G can be recovered at imaginary frequency nodes
-  EXPECT_LT(max_element(abs(ifops.coefs2vals(beta, gc) - g)), 1e-13);
+  EXPECT_LT(max_element(abs(ifops.coefs2vals(beta, gc) - g)), 2e-13);
 
   // Compute error in imaginary frequency
   auto gtru      = nda::matrix<dcomplex>(norb, norb);
