### 2023-10-28 721d9f0
[ghactions] Use libc++ for clang builds
