### 2023-11-07 861df93
added fmt, added unit test error printing
