+++ test/c++/imtime_ops.cpp
@@ -25,6 +25,7 @@
 #include <nda/nda.hpp>
 #include <cppdlr/cppdlr.hpp>
 #include <nda/gtest_tools.hpp>
+#include <fmt/format.h>
 
 using namespace cppdlr;
 using namespace nda;
@@ -123,6 +124,8 @@ TEST(imtime_ops, interp_matrix) {
   int nmaxtst = 10000; // # imag freq test points
 
   int norb = 2; // Orbital dimensions
+  
+  std::cout << fmt::format("eps = {:e}, Lambda = {:e}\n", eps, lambda);
 
   // Get DLR frequencies
   auto dlr_rf = build_dlr_rf(lambda, eps);
@@ -160,6 +163,7 @@ TEST(imtime_ops, interp_matrix) {
 
   EXPECT_LT(errlinf, 10 * eps);
   EXPECT_LT(errl2, eps);
+  std::cout << fmt::format("Imag time: L^2 err = {:e}, L^inf err = {:e}\n", errl2, errlinf);
 
   // Compute error in imaginary frequency
   auto ifops = imfreq_ops(lambda, dlr_rf, Fermion);
@@ -177,6 +181,7 @@ TEST(imtime_ops, interp_matrix) {
 
   EXPECT_LT(errlinf, 100 * eps);
   EXPECT_LT(errl2, eps);
+  std::cout << fmt::format("Imag freq: l^2 err = {:e}, L^inf err = {:e}\n", errl2, errlinf);
 }
 
 /**
@@ -192,6 +197,8 @@ TEST(imtime_ops, interp_matrix_complex) {
   int ntst    = 10000; // # imag time test points
 
   int norb = 2; // Orbital dimensions
+  
+  std::cout << fmt::format("eps = {:e}, Lambda = {:e}\n", eps, lambda);
 
   // Get DLR frequencies
   auto dlr_rf = build_dlr_rf(lambda, eps);
@@ -229,6 +236,7 @@ TEST(imtime_ops, interp_matrix_complex) {
 
   EXPECT_LT(errlinf, 10 * eps);
   EXPECT_LT(errl2, eps);
+  std::cout << fmt::format("Imag time: L^2 err = {:e}, L^inf err = {:e}\n", errl2, errlinf);
 }
 
 /**
@@ -242,6 +250,8 @@ TEST(imtime_ops, interp_scalar) {
 
   double beta = 1000;  // Inverse temperature
   int ntst    = 10000; // # imag time test points
+
+  std::cout << fmt::format("eps = {:e}, Lambda = {:e}\n", eps, lambda);
 
   // Get DLR frequencies
   auto dlr_rf = build_dlr_rf(lambda, eps);
@@ -277,6 +287,7 @@ TEST(imtime_ops, interp_scalar) {
 
   EXPECT_LT(errlinf, 10 * eps);
   EXPECT_LT(errl2, eps);
+  std::cout << fmt::format("Imag time: L^2 err = {:e}, L^inf err = {:e}\n", errl2, errlinf);
 
   // Test that constructing vector of evaluation at a point and then applying to
   // coefficients gives same result as direct evaluation method
@@ -298,6 +309,9 @@ TEST(imtime_ops, fit_matrix) {
   int ntst     = 10000; // # imag time test points
 
   int norb = 2; // Orbital dimensions
+
+  std::cout << fmt::format("eps = {:e}, Lambda = {:e}, noise = {:e}\n", eps, lambda, noise);
+
 
   // Get DLR frequencies
   auto dlr_rf = build_dlr_rf(lambda, eps);
@@ -331,6 +345,7 @@ TEST(imtime_ops, fit_matrix) {
 
[…]
