### 2023-11-11 047ace4
remove statistic option in real freq and im time [symmetrized im freq now broken]
