### 2023-11-11 d4e0e17
remove statistic option in real freq and im time [symmetrized im freq now broken]
