### 2023-11-11 6ea5f1d
pivrgs_sym does not allow middle row as pivot if odd # rows (for symmetrized bosonic im freq nodes)
