### 2023-11-11 3f517f2
implemented pivrgs_sym with rank specified; only case in which odd # rows allowed
