### 2023-11-11 c0ba9bb
n=0 added for symmetrized bosonic case, vals2coefs does overdetermined solve
