+++ c++/cppdlr/dlr_imfreq.cpp
@@ -15,6 +15,7 @@
 // Authors: Nils Wentzell, Jason Kaye
 
 #include "dlr_imfreq.hpp"
+#include "cppdlr/dlr_kernels.hpp"
 
 using namespace nda;
 
@@ -22,64 +23,35 @@ namespace cppdlr {
 
   imfreq_ops::imfreq_ops(double lambda, nda::vector_const_view<double> dlr_rf, statistic_t statistic, bool symmetrize)
      : lambda_(lambda), statistic(statistic), r(dlr_rf.size()), dlr_rf(dlr_rf) {
-
-    dlr_if    = nda::vector<int>(r);
-    cf2if     = nda::matrix<dcomplex>(r, r);
-    if2cf.lu  = nda::matrix<dcomplex>(r, r);
-    if2cf.piv = nda::vector<int>(r);
+    
+    // Get # DLR imaginary frequency nodes; for symmetrized bosonic case, this
+    // is DLR rank + 1, otherwise it is DLR rank
+    niom = (statistic == Boson && symmetrize) ? r + 1 : r;
+    dlr_if   = nda::vector<int>(niom);
+    cf2if    = nda::matrix<dcomplex>(niom, r);
 
     // Get analytic continuation kernel at DLR frequencies, up to imaginary
     // frequency cutoff
     auto nmax = fineparams(lambda).nmax;
     auto kmat = build_k_if(nmax, dlr_rf, statistic);
 
-    if (!(symmetrize && statistic == Boson)) { // Treat symmetrized bosonic case separately
+    // Pivoted Gram-Schmidt to obtain DLR imaginary frequency nodes
+    auto [q, norms, piv] = (symmetrize ? pivrgs_sym(kmat, niom) : pivrgs(kmat, 1e-100));
+    std::sort(piv.begin(), piv.end()); // Sort pivots in ascending order
+    for (int i = 0; i < niom; ++i) { dlr_if(i) = piv(i) - nmax; }
 
-      // Pivoted Gram-Schmidt to obtain DLR imaginary frequency nodes
-      auto [q, norms, piv] = (symmetrize ? pivrgs_sym(kmat, 1e-100) : pivrgs(kmat, 1e-100));
-      std::sort(piv.begin(), piv.end()); // Sort pivots in ascending order
-      for (int i = 0; i < r; ++i) { dlr_if(i) = piv(i) - nmax; }
-
-      // Obtain coefficients to imaginary frequency values transformation matrix
-      for (int i = 0; i < r; ++i) {
-        for (int j = 0; j < r; ++j) { cf2if(i, j) = kmat(piv(i), j); }
-      }
-
-    } else { // Symmetrized bosonic case: enforce n = 0 as DLR imaginary frequency node
-
-      auto kvec0 = nda::vector<dcomplex>(kmat(nmax, range(r))); // K at n = 0: K(0, om)
-
-      // Remove row of K matrix corresponding to n = 0 and column corresponding
-      // to omega=0
-      kmat(range(nmax, 2 * nmax), range(r)) = kmat(range(nmax + 1, 2 * nmax + 1), range(r));
-
-      // Pivoted Gram-Schmidt on rows of modified K matrix, augmented by vector
-      // K(0, om) which was removed, to obtain DLR imaginary frequency nodes
-      auto [q, norms, piv] = pivrgs_sym(kmat(range(2 * nmax), range(r)), kvec0, 1e-100);
-
-      std::sort(piv.begin(), piv.end());      // Sort pivots in ascending order
-      for (int i = 0; i < (r - 1) / 2; ++i) { // Fill in negative im freq nodes
-        dlr_if(i) = piv(i + 1) - 1 - nmax;    // Shift by 1 to obtain pivots of original matrix, not augmented matrix
-      }
[…]
