### 2023-11-12 72735b2
Correctly propagate fmt CMake targets
Fixes https://github.com/TRIQS/dft_tools/issues/245
