### 2023-11-12 b3c7cb7
Adjust test tolerances
Fixes https://github.com/flatironinstitute/cppdlr/issues/6
