+++ test/c++/imtime_ops.cpp
@@ -162,7 +162,7 @@ TEST(imtime_ops, interp_matrix) {
   errl2 = sqrt(errl2 / ntst);
 
   EXPECT_LT(errlinf, 10 * eps);
-  EXPECT_LT(errl2, eps);
+  EXPECT_LT(errl2, 2 * eps);
   std::cout << fmt::format("Imag time: L^2 err = {:e}, L^inf err = {:e}\n", errl2, errlinf);
 
   // Compute error in imaginary frequency
@@ -180,7 +180,7 @@ TEST(imtime_ops, interp_matrix) {
   errl2 = sqrt(errl2) / beta;
 
   EXPECT_LT(errlinf, 100 * eps);
-  EXPECT_LT(errl2, eps);
+  EXPECT_LT(errl2, 2 * eps);
   std::cout << fmt::format("Imag freq: l^2 err = {:e}, L^inf err = {:e}\n", errl2, errlinf);
 }
 
@@ -235,7 +235,7 @@ TEST(imtime_ops, interp_matrix_complex) {
   errl2 = sqrt(errl2 / ntst);
 
   EXPECT_LT(errlinf, 10 * eps);
-  EXPECT_LT(errl2, eps);
+  EXPECT_LT(errl2, 2 * eps);
   std::cout << fmt::format("Imag time: L^2 err = {:e}, L^inf err = {:e}\n", errl2, errlinf);
 }
 
@@ -529,10 +529,10 @@ TEST(imtime_ops, convolve_scalar_real) {
   errl2  = sqrt(errl2 / ntst);
   errtl2 = sqrt(errtl2 / ntst);
 
-  EXPECT_LT(errlinf, 10 * eps);
-  EXPECT_LT(errtlinf, 10 * eps);
-  EXPECT_LT(errl2, eps);
-  EXPECT_LT(errtl2, eps);
+  EXPECT_LT(errlinf, 25 * eps);
+  EXPECT_LT(errtlinf, 25 * eps);
+  EXPECT_LT(errl2, 4 * eps);
+  EXPECT_LT(errtl2, 3 * eps);
   std::cout << fmt::format("Ordinary convolution: L^inf err = {:e}, L^2 err = {:e}\n", errlinf, errl2);
   std::cout << fmt::format("Time-ordered convolution: L^inf err = {:e}, L^2 err = {:e}\n", errtlinf, errtl2);
 }
@@ -615,10 +615,10 @@ TEST(imtime_ops, convolve_scalar_cmplx) {
   errl2  = sqrt(errl2 / ntst);
   errtl2 = sqrt(errtl2 / ntst);
 
-  EXPECT_LT(errlinf, 10 * eps);
-  EXPECT_LT(errtlinf, 10 * eps);
-  EXPECT_LT(errl2, eps);
-  EXPECT_LT(errtl2, eps);
+  EXPECT_LT(errlinf, 13 * eps);
+  EXPECT_LT(errtlinf, 13 * eps);
+  EXPECT_LT(errl2, 2 * eps);
+  EXPECT_LT(errtl2, 2 * eps);
   std::cout << fmt::format("Ordinary convolution: L^inf err = {:e}, L^2 err = {:e}\n", errlinf, errl2);
   std::cout << fmt::format("Time-ordered convolution: L^inf err = {:e}, L^2 err = {:e}\n", errtlinf, errtl2);
 }
@@ -705,10 +705,10 @@ TEST(imtime_ops, convolve_matrix_real) {
   errl2  = sqrt(errl2 / ntst);
   errtl2 = sqrt(errtl2 / ntst);
 
-  EXPECT_LT(errlinf, 10 * eps);
-  EXPECT_LT(errl2, 2 * eps);
[…]
