+++ docs/main/doxygen/latex/dlr__imfreq_8cpp.tex
@@ -1,6 +1,7 @@
 \hypertarget{dlr__imfreq_8cpp}{}\doxysection{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/dlr\+\_\+imfreq.cpp File Reference}
 \label{dlr__imfreq_8cpp}\index{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/dlr\_imfreq.cpp@{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/dlr\_imfreq.cpp}}
 {\ttfamily \#include \char`\"{}dlr\+\_\+imfreq.\+hpp\char`\"{}}\newline
+{\ttfamily \#include \char`\"{}cppdlr/dlr\+\_\+kernels.\+hpp\char`\"{}}\newline
 \doxysubsection*{Namespaces}
 \begin{DoxyCompactItemize}
 \item 
