+++ docs/main/doxygen/latex/dlr__build_8hpp.tex
@@ -40,7 +40,7 @@ nda\+::matrix$<$ dcomplex $>$ \mbox{\hyperlink{namespacecppdlr_a988f80420d11cd80
 \begin{DoxyCompactList}\small\item\em Get imaginary frequency discretization of analytic continuation kernel. \end{DoxyCompactList}\item 
 std\+::tuple$<$ double, double $>$ \mbox{\hyperlink{namespacecppdlr_a48fb848c5a66d6141ad240d2e749aec9}{cppdlr\+::geterr\+\_\+k\+\_\+it}} (fineparams \&fine, nda\+::vector\+\_\+const\+\_\+view$<$ double $>$ t, nda\+::vector\+\_\+const\+\_\+view$<$ double $>$ om, nda\+::matrix\+\_\+const\+\_\+view$<$ double $>$ kmat)
 \begin{DoxyCompactList}\small\item\em Get error of fine composite Chebyshev discretization of analytic continuation kernel in imaginary time. \end{DoxyCompactList}\item 
-nda\+::vector$<$ double $>$ \mbox{\hyperlink{namespacecppdlr_a134a65b6f951a80cb99e41d52d89f0ed}{cppdlr\+::build\+\_\+dlr\+\_\+rf}} (double lambda, double eps, statistic\+\_\+t statistic, bool symmetrize)
+nda\+::vector$<$ double $>$ \mbox{\hyperlink{namespacecppdlr_a07f27d33c07f08031768f0844d14724e}{cppdlr\+::build\+\_\+dlr\+\_\+rf}} (double lambda, double eps, bool symmetrize)
 \begin{DoxyCompactList}\small\item\em Construct DLR basis by obtaining DLR frequencies. \end{DoxyCompactList}\item 
 nda\+::vector$<$ double $>$ \mbox{\hyperlink{namespacecppdlr_ace84f47b6cc94a562dfd5207791c27b0}{cppdlr\+::build\+\_\+dlr\+\_\+rf}} (double lambda, double eps)
 \begin{DoxyCompactList}\small\item\em Construct DLR basis by obtaining DLR frequencies. \end{DoxyCompactList}\end{DoxyCompactItemize}
