### 2024-01-23 bf579de
make complex copy of it2cf.lu in imtime_ops constructor to avoid on-the-fly copies
