+++ c++/cppdlr/dlr_imtime.hpp
@@ -32,7 +32,7 @@ namespace cppdlr {
   /**
   * Option for ordinary or time-ordered convolution 
   */
-   static constexpr bool ORDINARY = false, TIME_ORDERED = true;
+  static constexpr bool ORDINARY = false, TIME_ORDERED = true;
 
   /**
   * @class imtime_ops
@@ -65,8 +65,8 @@ namespace cppdlr {
     imtime_ops(double lambda, nda::vector_const_view<double> dlr_rf);
 
     imtime_ops(double lambda, nda::vector_const_view<double> dlr_rf, nda::vector_const_view<double> dlr_it, nda::matrix_const_view<double> cf2it,
-               nda::matrix_const_view<double> it2cf_lu, nda::vector_const_view<int> it2cf_piv)
-       : lambda_(lambda), r(dlr_rf.size()), dlr_rf(dlr_rf), dlr_it(dlr_it), cf2it(cf2it), it2cf{it2cf_lu, it2cf_piv} {};
+               nda::matrix_const_view<double> it2cf_lu, nda::matrix_const_view<dcomplex> it2cf_zlu, nda::vector_const_view<int> it2cf_piv)
+       : lambda_(lambda), r(dlr_rf.size()), dlr_rf(dlr_rf), dlr_it(dlr_it), cf2it(cf2it), it2cf{it2cf_lu, it2cf_zlu, it2cf_piv} {};
 
     imtime_ops() = default;
 
@@ -93,11 +93,10 @@ namespace cppdlr {
       auto gfv = nda::reshape(gf, r, g.size() / r);
 
       // Solve linear system (multiple right hand sides) to convert vals -> coeffs
-      if constexpr (nda::have_same_value_type_v<T, decltype(it2cf.lu)>) {
-        nda::lapack::getrs(it2cf.lu, gfv, it2cf.piv);
+      if constexpr (nda::is_complex_v<S>) { // getrs requires matrix and rhs to have same value type
+        nda::lapack::getrs(it2cf.zlu, gfv, it2cf.piv);
       } else {
-        // getrs requires matrix and rhs to have same value type
-        nda::lapack::getrs(nda::matrix<S>(it2cf.lu), gfv, it2cf.piv);
+        nda::lapack::getrs(it2cf.lu, gfv, it2cf.piv);
       }
 
       return gf;
@@ -256,7 +255,6 @@ namespace cppdlr {
         return gr;
       }
     }
-
 
     /** 
     * @brief Compute convolution of two imaginary time Green's functions
@@ -470,11 +468,11 @@ namespace cppdlr {
         fconv += matmul(cf2it, tmp2);
 
         // Then precompose with DLR grid values to DLR coefficients matrix
-        if constexpr (nda::have_same_value_type_v<T, decltype(it2cf.lu)>) {
-          nda::lapack::getrs(transpose(it2cf.lu), fconv, it2cf.piv); // Note: lapack effectively tranposes fconv by fortran reordering here
+        if constexpr (nda::is_complex_v<S>) {                         // getrs requires matrix and rhs to have same value type
+          nda::lapack::getrs(transpose(it2cf.zlu), fconv, it2cf.piv); // Note: lapack effectively tranposes fconv by fortran reordering here
         } else {
           // getrs requires matrix and rhs to have same value type
-          nda::lapack::getrs(transpose(nda::matrix<S>(it2cf.lu)), fconv, it2cf.piv);
+          nda::lapack::getrs(transpose(it2cf.lu), fconv, it2cf.piv);
         }
 
         return beta * fconv;
@@ -517,11 +515,11 @@ namespace cppdlr {
         }
 
         // Do the solve
[…]
