+++ docs/main/doxygen/latex/classcppdlr_1_1imtime__ops.tex
@@ -16,7 +16,7 @@ Class responsible for all DLR imaginary time operations, including building imag
 \begin{DoxyCompactList}\small\item\em Constructor for \mbox{\hyperlink{classcppdlr_1_1imtime__ops}{imtime\+\_\+ops}}. \end{DoxyCompactList}\item 
 \mbox{\hyperlink{classcppdlr_1_1imtime__ops_a0a1d4dfc2eb1ebc6fbd4cf6b9c13b2d7}{imtime\+\_\+ops}} (double \mbox{\hyperlink{classcppdlr_1_1imtime__ops_a401521dedc309ca12c1ec345da983c80}{lambda}}, nda\+::vector\+\_\+const\+\_\+view$<$ double $>$ dlr\+\_\+rf)
 \begin{DoxyCompactList}\small\item\em Constructor for \mbox{\hyperlink{classcppdlr_1_1imtime__ops}{imtime\+\_\+ops}}. \end{DoxyCompactList}\item 
-\mbox{\hyperlink{classcppdlr_1_1imtime__ops_a43a9b5fb6aeefad7d8ae6865329e69de}{imtime\+\_\+ops}} (double \mbox{\hyperlink{classcppdlr_1_1imtime__ops_a401521dedc309ca12c1ec345da983c80}{lambda}}, nda\+::vector\+\_\+const\+\_\+view$<$ double $>$ dlr\+\_\+rf, nda\+::vector\+\_\+const\+\_\+view$<$ double $>$ dlr\+\_\+it, nda\+::matrix\+\_\+const\+\_\+view$<$ double $>$ cf2it, nda\+::matrix\+\_\+const\+\_\+view$<$ double $>$ it2cf\+\_\+lu, nda\+::vector\+\_\+const\+\_\+view$<$ int $>$ it2cf\+\_\+piv)
+\mbox{\hyperlink{classcppdlr_1_1imtime__ops_ac44ddf86d5fd185dcf12add54362a6c0}{imtime\+\_\+ops}} (double \mbox{\hyperlink{classcppdlr_1_1imtime__ops_a401521dedc309ca12c1ec345da983c80}{lambda}}, nda\+::vector\+\_\+const\+\_\+view$<$ double $>$ dlr\+\_\+rf, nda\+::vector\+\_\+const\+\_\+view$<$ double $>$ dlr\+\_\+it, nda\+::matrix\+\_\+const\+\_\+view$<$ double $>$ cf2it, nda\+::matrix\+\_\+const\+\_\+view$<$ double $>$ it2cf\+\_\+lu, nda\+::matrix\+\_\+const\+\_\+view$<$ dcomplex $>$ it2cf\+\_\+zlu, nda\+::vector\+\_\+const\+\_\+view$<$ int $>$ it2cf\+\_\+piv)
 \item 
 \mbox{\hyperlink{classcppdlr_1_1imtime__ops_a8da78cf54d39d48d2be19ad1f146b03c}{imtime\+\_\+ops}} ()=default
 \item 
@@ -50,6 +50,8 @@ nda\+::matrix\+\_\+const\+\_\+view$<$ double $>$ \mbox{\hyperlink{classcppdlr_1_
 \begin{DoxyCompactList}\small\item\em Get transformation matrix from DLR coefficients to values at DLR imaginary time nodes. \end{DoxyCompactList}\item 
 nda\+::matrix\+\_\+const\+\_\+view$<$ double $>$ \mbox{\hyperlink{classcppdlr_1_1imtime__ops_af2b17e780b762dde6a25d8f8e844dab3}{get\+\_\+it2cf\+\_\+lu}} () const
[…]
