### 2024-01-29 #6297 @editorialbot
[REVIEW]: cppdlr: Imaginary time calculations using the discrete Lehmann representation
**Submitting author:** @jasonkaye (<a href="http://orcid.org/0000-0001-8045-6179">Jason Kaye</a>)
**Repository:** https://github.com/flatironinstitute/cppdlr/
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.2.0

@jarvist and @stevenrbrandt, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
