### 2024-02-01 831744c
Fix unitialized niom value in dlr_imfreq constructor
