### 2024-02-20 0e9888d
print warning for too small eps
