### 2024-02-22 6862d1e
method to compute tau inner prod of dlr expansions
