### 2024-03-05 8454355
Fix compiler warning, use abs from std namespace
