+++ joss/paper.bib
@@ -38,7 +38,7 @@ volume = {280},
 pages = {108458},
 year = {2022},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2022.108458},
+doi = {10.1016/j.cpc.2022.108458},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465522001771},
 author = {Jason Kaye and Kun Chen and Hugo U. R. Strand},
 keywords = {Many-body quantum physics, Imaginary time Green's functions, Low rank compression},
@@ -46,7 +46,7 @@ keywords = {Many-body quantum physics, Imaginary time Green's functions, Low ran
 
 @article{kaye23_eqdyson,
 author = {Kaye, Jason and Strand, Hugo U. R.},
-title = {A Fast Time Domain Solver for the Equilibrium Dyson Equation},
+title = {A Fast Time Domain Solver for the Equilibrium {D}yson Equation},
 year = {2023},
 _issue_date = {Aug 2023},
 publisher = {Springer-Verlag},
@@ -97,7 +97,7 @@ pages = "398-415",
 year = "2015",
 note = "",
 issn = "0010-4655",
-doi = "http://dx.doi.org/10.1016/j.cpc.2015.04.023",
+doi = "10.1016/j.cpc.2015.04.023",
 url = "http://www.sciencedirect.com/science/article/pii/S0010465515001666",
 author = "Olivier Parcollet and Michel Ferrero and Thomas Ayral and Hartmut Hafermann and Igor Krivenko and Laura Messio and Priyanka Seth"
 }
@@ -216,24 +216,35 @@ author = "Olivier Parcollet and Michel Ferrero and Thomas Ayral and Hartmut Hafe
   url = {https://link.aps.org/doi/10.1103/PhysRevB.106.L220502}
 }
 
-@misc{hou23,
-      title={Precursory {C}ooper Flow in Ultralow-Temperature Superconductors}, 
-      author={Pengcheng Hou and Xiansheng Cai and Tao Wang and Youjin Deng and Nikolay V. Prokof'ev and Boris V. Svistunov and Kun Chen},
-      year={2023},
-      eprint={2303.03624},
-      archivePrefix={arXiv},
-      primaryClass={cond-mat.supr-con}
+@article{hou24,
+  title = {Precursory {C}ooper flow in ultralow-temperature superconductors},
+  author = {Hou, Pengcheng and Cai, Xiansheng and Wang, Tao and Deng, Youjin and Prokof'ev, Nikolay V. and Svistunov, Boris V. and Chen, Kun},
+  journal = {Phys. Rev. Res.},
+  volume = {6},
+  issue = {1},
+  pages = {013099},
+  numpages = {10},
+  year = {2024},
+  month = {Jan},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevResearch.6.013099},
+  url = {https://link.aps.org/doi/10.1103/PhysRevResearch.6.013099}
 }
 
-@misc{johnston23,
-      title = {A comparative study of the superconductivity in the {H}olstein and optical {S}u-{S}chrieffer-{H}eeger models},
-      author = {{Tanjaroon Ly}, Andy and {Cohen-Stead}, Benjamin and {Malkaruge Costa}, Sohan and {Johnston}, Steven},
-      year = {2023},
-      eprint = {2307.10809},
-      archivePrefix = {arXiv},
-      primaryClass = {cond-mat.supr-con},
+@article{tanjaroonly23,
+  title = {Comparative study of the superconductivity in the {H}olstein and optical {Su-Schrieffer-Heeger} models},
[…]
