+++ joss/paper.bib
@@ -282,7 +282,7 @@ author = "Olivier Parcollet and Michel Ferrero and Thomas Ayral and Hartmut Hafe
   }
 
 @misc{cppdlr_git,
-	author = {Kaye, Jason and Wentzell, Nils and Strand, Hugo U. R.},
+	author = {Kaye, Jason and Strand, Hugo U. R., and Wentzell, Nils},
     journal = {GitHub repository},
   	publisher = {GitHub},
 	url = {https://github.com/flatironinstitute/cppdlr},
@@ -290,7 +290,7 @@ author = "Olivier Parcollet and Michel Ferrero and Thomas Ayral and Hartmut Hafe
 	year = {2023}}
 
 @misc{cppdlr_doc,
-	author = {Kaye, Jason and Wentzell, Nils and Strand, Hugo U. R.},
+	author = {Kaye, Jason and Strand, Hugo U. R., and Wentzell, Nils},
     journal = {GitHub-hosted documentation},
 	url = {https://flatironinstitute.github.io/cppdlr/},
 	title = {cppdlr: Imaginary time calculations using the discrete {L}ehmann representation},
