### 2024-04-04 d12a198
[ghactions] Synchronize build file with app4triqs, add caching using ccache
