### 2024-04-04 aced8e1
[ghactions] Synchronize build workflow with main branch
