### 2024-04-04 f7b1006
[cmake] Update top-level CMakeLists.txt file with latest app4triqs skeleton
