### 2024-04-23 9b72665
expose k_if_fermion and k_if_boson
