+++ c++/cppdlr/dlr_kernels.cpp
@@ -31,9 +31,7 @@ namespace cppdlr {
     }
   }
 
-  double k_it(double t, double om, double beta) {
-    return k_it(t, beta * om);
-  }
+  double k_it(double t, double om, double beta) { return k_it(t, beta * om); }
 
   double k_it_abs(double t, double om) {
 
@@ -61,8 +59,6 @@ namespace cppdlr {
       return k_if_boson(n, om);
   }
 
-  std::complex<double> k_if(int n, double om, statistic_t statistic, double beta) {
-    return beta * k_if(n, beta * om, statistic);
-  }
-    
+  std::complex<double> k_if(int n, double om, statistic_t statistic, double beta) { return beta * k_if(n, beta * om, statistic); }
+
 } // namespace cppdlr
