+++ docs/main/doxygen/latex/dlr__kernels_8cpp.tex
@@ -16,10 +16,10 @@ double \mbox{\hyperlink{namespacecppdlr_a3cfc4afda9cacac00e93425fe9d45a93}{cppdl
 \begin{DoxyCompactList}\small\item\em Evaluate analytic continuation kernel in imaginary time (relative time format) for a given inverse temperature. \end{DoxyCompactList}\item 
 double \mbox{\hyperlink{namespacecppdlr_a7dfaf907221eb1768418258421af703f}{cppdlr\+::k\+\_\+it\+\_\+abs}} (double t, double om)
 \begin{DoxyCompactList}\small\item\em Evaluate analytic continuation kernel in imaginary time (absolute time format) using dimensionless variables (beta = 1) \end{DoxyCompactList}\item 
-std\+::complex$<$ double $>$ \mbox{\hyperlink{namespacecppdlr_abbfc2f4ecdd39b95d4419a4be35909a4}{cppdlr\+::k\+\_\+if\+\_\+fermion}} (int n, double om)
-\item 
-std\+::complex$<$ double $>$ \mbox{\hyperlink{namespacecppdlr_a003438d56c2ca900d2f08af9f2ab424d}{cppdlr\+::k\+\_\+if\+\_\+boson}} (int n, double om)
-\item 
+std\+::complex$<$ double $>$ \mbox{\hyperlink{namespacecppdlr_a66270b210c46b39737f361467f645be7}{cppdlr\+::k\+\_\+if\+\_\+fermion}} (int n, double om)
+\begin{DoxyCompactList}\small\item\em Evaluate fermionic analytic continuation kernel in imaginary frequency using dimensionless variables (beta = 1) \end{DoxyCompactList}\item 
+std\+::complex$<$ double $>$ \mbox{\hyperlink{namespacecppdlr_ab73f52d9e0209e584d1210d419f39f8c}{cppdlr\+::k\+\_\+if\+\_\+boson}} (int n, double om)
+\begin{DoxyCompactList}\small\item\em Evaluate bosonic analytic continuation kernel in imaginary frequency using dimensionless variables (beta = 1) \end{DoxyCompactList}\item 
 std\+::complex$<$ double $>$ \mbox{\hyperlink{namespacecppdlr_af8f77e12b089b5ddb5b5b3cda18cf623}{cppdlr\+::k\+\_\+if}} (int n, double om, statistic\+\_\+t statistic)
 \begin{DoxyCompactList}\small\item\em Evaluate analytic continuation kernel in imaginary frequency using dimensionless variables (beta = 1) \end{DoxyCompactList}\item 
 std\+::complex$<$ double $>$ \mbox{\hyperlink{namespacecppdlr_a672098c598c55c8f9da60cd194a741ea}{cppdlr\+::k\+\_\+if}} (int n, double om, statistic\+\_\+t statistic, double beta)
