### 2024-05-26 ed3b5b9
[cmake] Bump version number to 1.1.0
