### 2024-05-30 2f6c824
Enable documentation builds for 1.1.x
