+++ .github/workflows/build.yml
@@ -113,7 +113,7 @@ jobs:
         key: ccache-${{ matrix.os }}-${{ matrix.cc }}-${{ github.run_id }}
 
     - name: Deploy documentation to website
-      if: matrix.doc == 'ON' && github.ref == 'refs/heads/main'
+      if: matrix.doc == 'ON' && github.ref == 'refs/heads/1.1.x'
       uses: JamesIves/github-pages-deploy-action@v4
       with:
         folder: build/doc/html
