+++ .github/workflows/build.yml
@@ -2,9 +2,9 @@ name: build
 
 on:
   push:
-    branches: [ main ]
+    branches: [ 1.1.x ]
   pull_request:
-    branches: [ main ]
+    branches: [ 1.1.x ]
   workflow_call:
   workflow_dispatch:
 
