### 2024-06-06 e8fc65b
allow imtime_ops::inner_prod to return double
