### 2024-06-18 c87f567
expose inner product matrix of imtime_ops
