### 2024-06-18 d0bfeea
added itops.vals2coefs option to apply transpose of operation, w/ unit test
