### 2024-06-18 7f7074a
inline documentation typo fix
