### 2024-06-20 e22d10e
vals2coefs documentation
