### 2024-07-01 #6297 @stevenrbrandt
OK, I'm getting started and walking my way through the checklist. I don't see anything that obviously fits the definition "A statement of need" in the docs. To me, a statement of need should probably have a title like "why this package was created" or something of that sort.
