### 2024-07-01 #6297 @stevenrbrandt
There seems to be a single example: https://flatironinstitute.github.io/cppdlr/main/examples.html

I was able to compile and run it with ease, and it produced a number of data files. I think it would be nice to see the scripts used to create the plots on the examples page from the data.
