### 2024-07-01 #6297 @stevenrbrandt
I am also having trouble identifying "community guidelines." All I see is that, to report issues, they authors point people to the github issue tracker.
