### 2024-07-01 #6297 @stevenrbrandt
I should probably add that the science the authors are doing is barely comprehensible to me as it's totally outside my field.

Again, I'm not sure what the requirements are for the statement of need, but one thing I would like to see is a high-level list of scientific questions / projects this code is used for, e.g. analysis of results from the LHC or something. Maybe some text borrowed from a community outreach event.
