### 2024-07-17 #6297 @stevenrbrandt
@jarvist @kyleniemeyer I have a few boxes that I haven't checked.

I do not see a "statement of need" in the docs. It's in the paper, though.

Automated tests look like they just test that the code builds, not that any result is repeatable. Is that sufficient?

While there are community guidelines, they just point users at the github issues link. Is that sufficient?

Thanks.
