### 2024-07-17 #6297 @jasonkaye
@stevenrbrandt Thanks for going through this carefully.

- For the statement of need, I think that for our users, the [Background page](https://flatironinstitute.github.io/cppdlr/latest/background.html) of our website should serve this purpose. However, if this is not sufficient, we could discuss what else needs to be added.
- We have automated unit tests which test essentially all aspects of the code, including reproducing practical results. These can be found in the test/c++ folder, and should run with `make test`. For example, test/c++/imtime_ops.cpp and test/c++/imfreq_ops.cpp contain tests for many of the main user-facing functions provided by the library, with realistic examples checking that user-specified error tolerances are satisfied.
