### 2024-07-22 #6297 @stevenrbrandt
@jasonkaye I really liked the statement of need in your paper much better. It is more of the form "here is the functionality we provide" and "here are other libraries which provide similar functionality." If you could make it a bit more like that, I think that would be great.
