### 2024-07-22 #6297 @stevenrbrandt
@jasonkaye also, it would be nice if for "community guidelines" you provided some guidance for people who want to make a PR.
