+++ doc/issues.rst
@@ -6,5 +6,7 @@ Issues and contributions
 Please report all bugs, as well as new feature requests, on the `GitHub issues page <https://github.com/flatironinstitute/cppdlr/issues>`_.
 
 If you wish to contribute new features to ``cppdlr``, please do so by creating a
-`Github pull request <https://github.com/flatironinstitute/cppdlr/pulls>`_. For
-major changes or new features, we recommend coordinating with the developers of ``cppdlr``.
+`Github pull request <https://github.com/flatironinstitute/cppdlr/pulls>`_,
+using the "main" branch, which serves as a development branch of ``cppdlr``, as
+the target branch. For major changes or new features, we recommend coordinating
+with the developers of ``cppdlr``.
