### 2024-07-22 #6297 @jasonkaye
@stevenrbrandt Thanks for the suggestions. We have added a paragraph on the [front page](https://flatironinstitute.github.io/cppdlr/main/index.html) of the documentation website which includes some of the points you mentioned. We have also renamed the "Issues" page to "[Issues and contributions](https://flatironinstitute.github.io/cppdlr/main/issues.html)", and included a brief statement on making pull requests.

Note that this was done in the [main (development) branch website](https://flatironinstitute.github.io/cppdlr/main/index.html), which is referenced and linked to in the github repo readme. It will be included in the stable branch website with our next release.
