### 2024-07-24 1a3c8f5
[h5] fix bwd compat for new field it2cf_zlu
