+++ c++/cppdlr/dlr_imtime.hpp
@@ -857,7 +857,8 @@ namespace cppdlr {
       auto it        = h5::read<nda::vector<double>>(gr, "it");
       auto cf2it_    = h5::read<nda::matrix<double>>(gr, "cf2it");
       auto it2cf_lu  = h5::read<nda::matrix<double>>(gr, "it2cf_lu");
-      auto it2cf_zlu = h5::read<nda::matrix<dcomplex>>(gr, "it2cf_zlu");
+      auto it2cf_zlu = nda::matrix<dcomplex>(it2cf_lu);
+      h5::try_read(gr, "it2cf_zlu", it2cf_zlu);
       auto it2cf_piv = h5::read<nda::vector<int>>(gr, "it2cf_piv");
 
       m = imtime_ops(lambda, rf, it, cf2it_, it2cf_lu, it2cf_zlu, it2cf_piv);
