### 2024-07-24 8c884cf
simplified it2cf_zlu treatment in h5_read
