### 2024-07-25 6ffd337
[h5] simplified it2cf_zlu treatment in h5_read for bwd compat w/ 1.0.x (#11)
* [h5] fix bwd compat for new field it2cf_zlu

* simplified it2cf_zlu treatment in h5_read

---------

Co-authored-by: Jason Kaye <jkaye@flatironinstitute.org>
