+++ docs/main/_sources/issues.rst.txt
@@ -1,12 +1,7 @@
 .. _issues:
 
-Issues and contributions
-************************
+Reporting issues
+****************
 
-Please report all bugs, as well as new feature requests, on the `GitHub issues page <https://github.com/flatironinstitute/cppdlr/issues>`_.
-
-If you wish to contribute new features to ``cppdlr``, please do so by creating a
-`Github pull request <https://github.com/flatironinstitute/cppdlr/pulls>`_,
-using the "main" branch, which serves as a development branch of ``cppdlr``, as
-the target branch. For major changes or new features, we recommend coordinating
-with the developers of ``cppdlr``.
+Please report all problems and bugs on the GitHub issues page
+`<https://github.com/flatironinstitute/cppdlr/issues>`_.
