### 2024-07-26 #6297 @jarvist
OK! Other than just hanging on making sure the 'community guidelines' are updated, I think we are all good.

Less usual to see a C++ scientific code today, but this one looked quite nicely and cleanly written, using modern C++ templates approach. I thought the abstraction to the `NDA` sub-library was interesting and quite nice, but then I wondered whether it would make sense for it to for-ever live in this repository.

Code comments were quite light in simplistic parts, but then went into a nice level of detail in some of the more tricky bits of physics, which felt like the right approach.

Codebase is compact, but I think this work definitely makes the JOSS threshold for scholarly effort as its very hard to get the numerics of these things correct, and it's certainly something that will accelerate the development of codes which make use of the discrete Lehmann representation.

It would have been nice to see a bit more description of how these codes compare to the Python, Julia and previous C (?) libraries that do the same. The motivation for a modern C++ code is absolutely fine though.
In an ideal world it would have been nice to have some more examples, but the one that you did have was very nicely documented with both the physics & then the code side by side.
