+++ docs/1.1.x/doxygen/latex/utils_8hpp.tex
@@ -1,0 +1,57 @@
+\hypertarget{utils_8hpp}{}\doxysection{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/utils.hpp File Reference}
+\label{utils_8hpp}\index{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/utils.hpp@{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/utils.hpp}}
+{\ttfamily \#include \char`\"{}nda/concepts.\+hpp\char`\"{}}\newline
+{\ttfamily \#include $<$nda/nda.\+hpp$>$}\newline
+{\ttfamily \#include $<$nda/blas.\+hpp$>$}\newline
+\doxysubsection*{Classes}
+\begin{DoxyCompactItemize}
+\item 
+class \mbox{\hyperlink{classcppdlr_1_1barycheb}{cppdlr\+::barycheb}}
+\item 
+class \mbox{\hyperlink{classcppdlr_1_1baryleg}{cppdlr\+::baryleg}}
+\item 
+struct \mbox{\hyperlink{structcppdlr_1_1make__common__helper}{cppdlr\+::make\+\_\+common\+\_\+helper$<$ T, S1, S2 $>$}}
+\begin{DoxyCompactList}\small\item\em Get type of given nda Memory\+Array with scalar value type replaced by common type of two given types (real if both are real, complex otherwise) \end{DoxyCompactList}\end{DoxyCompactItemize}
+\doxysubsection*{Namespaces}
+\begin{DoxyCompactItemize}
+\item 
+ \mbox{\hyperlink{namespacecppdlr}{cppdlr}}
+\end{DoxyCompactItemize}
+\doxysubsection*{Typedefs}
+\begin{DoxyCompactItemize}
+\item 
+{\footnotesize template$<$nda\+::\+Memory\+Array T$>$ }\\using \mbox{\hyperlink{namespacecppdlr_a85cd7bdb1fd649296322df16d329ca27}{cppdlr\+::make\+\_\+real\+\_\+t}} = decltype(make\+\_\+regular(real(std\+::declval$<$ T $>$())))
+\begin{DoxyCompactList}\small\item\em Get real-\/valued type corresponding to type of given nda Memory\+Array. \end{DoxyCompactList}\item 
+{\footnotesize template$<$nda\+::\+Memory\+Array T$>$ }\\using \mbox{\hyperlink{namespacecppdlr_af7426dd83d2c831f6ae3fa9b8e6172ad}{cppdlr\+::make\+\_\+cplx\+\_\+t}} = decltype(make\+\_\+regular(std\+::declval$<$ T $>$() $\ast$1i))
+\begin{DoxyCompactList}\small\item\em Get complex-\/valued type corresponding to type of given nda Memory\+Array. \end{DoxyCompactList}\item 
+{\footnotesize template$<$nda\+::\+Memory\+Array T, nda\+::\+Scalar S1, nda\+::\+Scalar S2$>$ }\\using \mbox{\hyperlink{namespacecppdlr_ac23dfec829ec6ed8deec469faeea3981}{cppdlr\+::make\+\_\+common\+\_\+t}} = typename make\+\_\+common\+\_\+helper$<$ T, S1, S2 $>$\+::type
+\end{DoxyCompactItemize}
+\doxysubsection*{Functions}
+\begin{DoxyCompactItemize}
+\item 
[…]
