+++ docs/1.1.x/doxygen/latex/dlr__imtime_8hpp.tex
@@ -1,0 +1,19 @@
+\hypertarget{dlr__imtime_8hpp}{}\doxysection{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/dlr\+\_\+imtime.hpp File Reference}
+\label{dlr__imtime_8hpp}\index{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/dlr\_imtime.hpp@{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/dlr\_imtime.hpp}}
+{\ttfamily \#include $<$nda/nda.\+hpp$>$}\newline
+{\ttfamily \#include \char`\"{}cppdlr/dlr\+\_\+kernels.\+hpp\char`\"{}}\newline
+{\ttfamily \#include \char`\"{}cppdlr/dlr\+\_\+build.\+hpp\char`\"{}}\newline
+{\ttfamily \#include \char`\"{}cppdlr/utils.\+hpp\char`\"{}}\newline
+{\ttfamily \#include \char`\"{}nda/clef/clef.\+hpp\char`\"{}}\newline
+{\ttfamily \#include $<$h5/h5.\+hpp$>$}\newline
+{\ttfamily \#include $<$nda/h5.\+hpp$>$}\newline
+\doxysubsection*{Classes}
+\begin{DoxyCompactItemize}
+\item 
+class \mbox{\hyperlink{classcppdlr_1_1imtime__ops}{cppdlr\+::imtime\+\_\+ops}}
+\begin{DoxyCompactList}\small\item\em Class responsible for all DLR imaginary time operations, including building imaginary time grid and transformations. \end{DoxyCompactList}\end{DoxyCompactItemize}
+\doxysubsection*{Namespaces}
+\begin{DoxyCompactItemize}
+\item 
+ \mbox{\hyperlink{namespacecppdlr}{cppdlr}}
+\end{DoxyCompactItemize}
