+++ docs/1.1.x/_sources/issues.rst.txt
@@ -1,0 +1,7 @@
+.. _issues:
+
+Reporting issues
+****************
+
+Please report all problems and bugs on the GitHub issues page
+`<https://github.com/flatironinstitute/cppdlr/issues>`_.

[…]
