### 2024-07-26 #6297 @jasonkaye
@jarvist Thanks for your comments. Thank you for catching the website error---this was caused by a bug related to the deployment of the docs to the website, and is now fixed. You should now be able to see the changes on the [main branch website](https://flatironinstitute.github.io/cppdlr/main/index.html), which will eventually make it to the stable branch.

We plan to add more examples as the need arises. The example we have will already cover probably 90% of user needs. As we describe on the [examples page](https://flatironinstitute.github.io/cppdlr/main/examples.html#list-of-other-cppdlr-capabilities), there are many other examples in the form of unit tests, and we give a map from features to the relevant unit tests on that page. Those unit tests themselves are also reasonably well documented, though not explained in as much detail as the main example.
