### 2024-07-30 #6297 @kyleniemeyer
I can see the change on the docs website regarding community guidelines, which I believe was @jarvist's only blocking issue - and so the reviews are now complete for this submission!
