### 2024-08-01 8844723
v1.2.0 changelog draft
