### 2024-08-02 0f2e7b3
[ghactions] Enable ghactions builds for release branches
