+++ .github/workflows/build.yml
@@ -2,9 +2,9 @@ name: build
 
 on:
   push:
-    branches: [ main ]
+    branches: [ main, '[0-9]+.[0-9]+.x' ]
   pull_request:
-    branches: [ main ]
+    branches: [ main, '[0-9]+.[0-9]+.x' ]
   workflow_call:
   workflow_dispatch:
 
