+++ .github/workflows/build.yml
@@ -113,9 +113,9 @@ jobs:
         key: ccache-${{ matrix.os }}-${{ matrix.cc }}-${{ github.run_id }}
 
     - name: Deploy documentation to website
-      if: matrix.doc == 'ON' && github.ref == 'refs/heads/main'
+      if: matrix.doc == 'ON' && github.event_name == 'push'
       uses: JamesIves/github-pages-deploy-action@v4
       with:
         folder: build/doc/html
         branch: github.io
-        target-folder: docs/main
+        target-folder: docs/${{ github.ref_name }}
