+++ docs/1.2.x/doxygen/latex/dlr__kernels_8hpp.tex
@@ -1,0 +1,32 @@
+\hypertarget{dlr__kernels_8hpp}{}\doxysection{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/dlr\+\_\+kernels.hpp File Reference}
+\label{dlr__kernels_8hpp}\index{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/dlr\_kernels.hpp@{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/dlr\_kernels.hpp}}
+{\ttfamily \#include $<$complex$>$}\newline
+\doxysubsection*{Namespaces}
+\begin{DoxyCompactItemize}
+\item 
+ \mbox{\hyperlink{namespacecppdlr}{cppdlr}}
+\end{DoxyCompactItemize}
+\doxysubsection*{Enumerations}
+\begin{DoxyCompactItemize}
+\item 
+enum \mbox{\hyperlink{namespacecppdlr_ad0a7e33d9a03f4986937032c25ee7435}{cppdlr\+::statistic\+\_\+t}} \{ \mbox{\hyperlink{namespacecppdlr_ad0a7e33d9a03f4986937032c25ee7435a6a57173d6b5be36c6815ab715152716d}{cppdlr\+::\+Boson}} = 0
+, \mbox{\hyperlink{namespacecppdlr_ad0a7e33d9a03f4986937032c25ee7435ac01c1699bb87ec8268ee33f7f20d84ee}{cppdlr\+::\+Fermion}} = 1
+ \}
+\end{DoxyCompactItemize}
+\doxysubsection*{Functions}
+\begin{DoxyCompactItemize}
+\item 
+double \mbox{\hyperlink{namespacecppdlr_afc9726e02898098dfff8448f02f5b3c6}{cppdlr\+::k\+\_\+it}} (double t, double om)
+\begin{DoxyCompactList}\small\item\em Evaluate analytic continuation kernel in imaginary time (relative time format) using dimensionless variables (beta = 1) \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{namespacecppdlr_a3cfc4afda9cacac00e93425fe9d45a93}{cppdlr\+::k\+\_\+it}} (double t, double om, double beta)
+\begin{DoxyCompactList}\small\item\em Evaluate analytic continuation kernel in imaginary time (relative time format) for a given inverse temperature. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{namespacecppdlr_a7dfaf907221eb1768418258421af703f}{cppdlr\+::k\+\_\+it\+\_\+abs}} (double t, double om)
+\begin{DoxyCompactList}\small\item\em Evaluate analytic continuation kernel in imaginary time (absolute time format) using dimensionless variables (beta = 1) \end{DoxyCompactList}\item 
+std\+::complex$<$ double $>$ \mbox{\hyperlink{namespacecppdlr_a66270b210c46b39737f361467f645be7}{cppdlr\+::k\+\_\+if\+\_\+fermion}} (int n, double om)
+\begin{DoxyCompactList}\small\item\em Evaluate fermionic analytic continuation kernel in imaginary frequency using dimensionless variables (beta = 1) \end{DoxyCompactList}\item 
+std\+::complex$<$ double $>$ \mbox{\hyperlink{namespacecppdlr_ab73f52d9e0209e584d1210d419f39f8c}{cppdlr\+::k\+\_\+if\+\_\+boson}} (int n, double om)
+\begin{DoxyCompactList}\small\item\em Evaluate bosonic analytic continuation kernel in imaginary frequency using dimensionless variables (beta = 1) \end{DoxyCompactList}\item 
[…]
