+++ docs/1.2.x/doxygen/latex/classcppdlr_1_1imfreq__ops.tex
@@ -1,0 +1,382 @@
+\hypertarget{classcppdlr_1_1imfreq__ops}{}\doxysection{cppdlr\+::imfreq\+\_\+ops Class Reference}
+\label{classcppdlr_1_1imfreq__ops}\index{cppdlr::imfreq\_ops@{cppdlr::imfreq\_ops}}
+
+
+Class responsible for all DLR imaginary frequency operations, including building imaginary frequency grid and transformations.  
+
+
+
+
+{\ttfamily \#include $<$dlr\+\_\+imfreq.\+hpp$>$}
+
+\doxysubsection*{Public Member Functions}
+\begin{DoxyCompactItemize}
+\item 
+\mbox{\hyperlink{classcppdlr_1_1imfreq__ops_a2296d91d2652b85462653f6147d95065}{imfreq\+\_\+ops}} (double \mbox{\hyperlink{classcppdlr_1_1imfreq__ops_a0b95a2d9da812c7f894c16d9ab0de27c}{lambda}}, nda\+::vector\+\_\+const\+\_\+view$<$ double $>$ dlr\+\_\+rf, \mbox{\hyperlink{namespacecppdlr_ad0a7e33d9a03f4986937032c25ee7435}{statistic\+\_\+t}} statistic, bool symmetrize=false)
+\begin{DoxyCompactList}\small\item\em Constructor for \mbox{\hyperlink{classcppdlr_1_1imfreq__ops}{imfreq\+\_\+ops}}. \end{DoxyCompactList}\item 
+\mbox{\hyperlink{classcppdlr_1_1imfreq__ops_afc68ce6e34ea9a23e763f9b1919ea573}{imfreq\+\_\+ops}} (double \mbox{\hyperlink{classcppdlr_1_1imfreq__ops_a0b95a2d9da812c7f894c16d9ab0de27c}{lambda}}, nda\+::vector\+\_\+const\+\_\+view$<$ double $>$ dlr\+\_\+rf, \mbox{\hyperlink{namespacecppdlr_ad0a7e33d9a03f4986937032c25ee7435}{statistic\+\_\+t}} statistic, nda\+::vector\+\_\+const\+\_\+view$<$ int $>$ dlr\+\_\+if, nda\+::matrix\+\_\+const\+\_\+view$<$ nda\+::dcomplex $>$ cf2if, nda\+::matrix\+\_\+const\+\_\+view$<$ nda\+::dcomplex $>$ if2cf\+\_\+lu, nda\+::vector\+\_\+const\+\_\+view$<$ int $>$ if2cf\+\_\+piv)
+\item 
+\mbox{\hyperlink{classcppdlr_1_1imfreq__ops_aec1f430ef8442851ca0003bc89620feb}{imfreq\+\_\+ops}} ()=default
+\item 
+{\footnotesize template$<$nda\+::\+Memory\+Array T$>$ }\\T\+::regular\+\_\+type \mbox{\hyperlink{classcppdlr_1_1imfreq__ops_a88c5359e14d8dd6c764a34aa67bbc343}{vals2coefs}} (double beta, T const \&g) const
+\begin{DoxyCompactList}\small\item\em Transform values of Green\textquotesingle{}s function G on DLR imaginary frequency grid to DLR coefficients. \end{DoxyCompactList}\item 
+{\footnotesize template$<$nda\+::\+Memory\+Array T$>$ }\\T\+::regular\+\_\+type \mbox{\hyperlink{classcppdlr_1_1imfreq__ops_ab0c49ae76da53c29c1e48c46ca6a2ded}{vals2coefs}} (T const \&g) const
+\item 
[…]
