+++ docs/1.2.x/doxygen/latex/dlr__imtime_8cpp.tex
@@ -1,0 +1,11 @@
+\hypertarget{dlr__imtime_8cpp}{}\doxysection{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/dlr\+\_\+imtime.cpp File Reference}
+\label{dlr__imtime_8cpp}\index{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/dlr\_imtime.cpp@{/home/runner/work/cppdlr/cppdlr/c++/cppdlr/dlr\_imtime.cpp}}
+{\ttfamily \#include \char`\"{}dlr\+\_\+imtime.\+hpp\char`\"{}}\newline
+{\ttfamily \#include \char`\"{}cppdlr/dlr\+\_\+kernels.\+hpp\char`\"{}}\newline
+{\ttfamily \#include \char`\"{}dlr\+\_\+build.\+hpp\char`\"{}}\newline
+{\ttfamily \#include \char`\"{}utils.\+hpp\char`\"{}}\newline
+\doxysubsection*{Namespaces}
+\begin{DoxyCompactItemize}
+\item 
+ \mbox{\hyperlink{namespacecppdlr}{cppdlr}}
+\end{DoxyCompactItemize}

[…]
