### 2024-08-02 073d9f0
[cmake] Fix issue with Doxygen output directory
