+++ .github/workflows/build.yml
@@ -2,9 +2,9 @@ name: build
 
 on:
   push:
-    branches: [ 1.1.x ]
+    branches: [ main, '[0-9]+.[0-9]+.x' ]
   pull_request:
-    branches: [ 1.1.x ]
+    branches: [ main, '[0-9]+.[0-9]+.x' ]
   workflow_call:
   workflow_dispatch:
 
@@ -113,9 +113,9 @@ jobs:
         key: ccache-${{ matrix.os }}-${{ matrix.cc }}-${{ github.run_id }}
 
     - name: Deploy documentation to website
-      if: matrix.doc == 'ON' && github.ref == 'refs/heads/1.1.x'
+      if: matrix.doc == 'ON' && github.event_name == 'push'
       uses: JamesIves/github-pages-deploy-action@v4
       with:
         folder: build/doc/html
         branch: github.io
-        target-folder: docs/1.1.x
+        target-folder: docs/${{ github.ref_name }}
