### 2024-08-02 c16575c
[cmake] Fix issue with Doxygen output directory
