### 2024-08-02 5d8d0e2
[doc] Incorporate Version 1.1.1 into changelog
