+++ doc/ChangeLog.md
@@ -12,15 +12,28 @@ We thank all contributors: Jason Kaye, Hugo U. R. Strand, Nils Wentzell
 * Added imtime_ops::get_ipmat function to expose matrix of imaginary time inner product
 * Implemented transpose of values -> coefficients operation in imtime_ops::vals2coefs 
 
-### Bug fixes
-* Fixed backwards compatibility issue w/ 1.0.x for h5 reading of it2cf_zlu object in imtime_ops 
-
 ### Optimizations
 * imtime_ops::inner_prod function can return double, rather than always complex double
 
 ### Other changes
 * Test tolerance adjustments
 * Minor documentation updates
+
+
+## Version 1.1.1
+
+cppdlr version 1.1.1 is a patch-release that includes
+cmake build system improvements, continuous integration improvements,
+and a backward compatibilty fix.
+
+We thank all contributors: Hugo U.R. Strand, Nils Wentzell
+
+Find below an itemized list of changes in this release.
+
+* Synchronize ghactions workflow with main
+* Use nda 1.3.x branch by default
+* Fix issue with Doxygen output directory
+* simplified it2cf_zlu treatment in h5_read for bwd compat
 
 
 ## Version 1.1.0
