### 2024-08-02 1c3992e
[doc] Add Version 1.1.1 to changelog
