+++ joss/paper.bib
@@ -72,6 +72,16 @@ _keywords = {81-10, 81T18, 81-08, 81S40, Nonlinear Volterra integral equations, 
       primaryClass={cond-mat.str-el}
 }
 
+@misc{kiese24,
+      title={Discrete Lehmann representation of three-point functions}, 
+      author={Dominik Kiese and Hugo U. R. Strand and Kun Chen and Nils Wentzell
+      and Olivier Parcollet and Jason Kaye},
+      year={2024},
+      eprint={2405.06716},
+      archivePrefix={arXiv},
+      primaryClass={physics.comp-ph}
+}
+
 @article{sheng23,
   title = {Low-rank {G}reen's function representations applied to dynamical mean-field theory},
   author = {Sheng, Nan and Hampel, Alexander and Beck, Sophie and Parcollet, Olivier and Wentzell, Nils and Kaye, Jason and Chen, Kun},
