+++ joss/paper.bib
@@ -65,7 +65,7 @@ _keywords = {81-10, 81T18, 81-08, 81S40, Nonlinear Volterra integral equations, 
 
 @misc{kaye23_diagrams,
       title={Decomposing imaginary time {F}eynman diagrams using separable basis functions: Anderson impurity model strong coupling expansion}, 
-      author={Jason Kaye and Hugo U. R. Strand and Denis Gole{\v z}},
+      author={Jason Kaye and Zhen Huang and Hugo U. R. Strand and Denis Gole{\v z}},
       year={2023},
       eprint={2307.08566},
       archivePrefix={arXiv},
