### 2024-08-06 #6297 @jasonkaye
@kyleniemeyer My apologies, but we needed to make a couple of small edits to affiliations and references, which we have pushed to our repo---see also the generated pdf above. If you could regenerate the final proof with these changes, I can take a final read over and confirm. Thanks.
