### 2024-08-06 #6297 @jasonkaye
> Congratulations @jasonkaye on your article's publication in JOSS! Thanks for your patience throughout this process.
>
> Many thanks to @jarvist and @stevenrbrandt for reviewing this submission.

Great, thanks @kyleniemeyer for your help.
